requirements.txt
fastapi==0.115.6
numpy==2.2.0
python-dotenv==1.0.1
uvicorn==0.34.0
